requirements.txt
appdirs==1.4.3
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-import-export==2.1.0
django-mathfilters==1.0.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
Naked==0.1.31
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.5
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.6.1
pygount==1.2.1
Pyphen==0.9.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
shellescape==3.8.1
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
tinycss2==1.0.2
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0